backend/requirements.txt
# backend/requirements.txt
<<<<<<< HEAD
fastapi[all]==0.129.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.36
alembic==1.13.1
psycopg2-binary>=2.9.10
pydantic>=2.6.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
argon2-cffi==23.1.0
python-multipart==0.0.6
redis==5.0.1
python-magic==0.4.27
cryptography>=44.0.0
anthropic==0.79.0
httpx==0.26.0
python-dotenv==1.0.0
requests>=2.32.4
=======
fastapi[all]==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
pydantic>=2.6.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
redis==5.0.1
python-magic==0.4.27
cryptography==42.0.0
anthropic==0.8.1
httpx==0.26.0
python-dotenv==1.0.0
requests>=2.31.0
>>>>>>> 2b6c4e0f3e2bdec671123c59cab390bd0dde93d7
google-auth==2.26.2
# Email
aiosmtplib==3.0.1  # Async SMTP client
email-validator==2.1.0  # Email validation
jinja2==3.1.3  # Email templates
# User agent parsing
user-agents==2.2.0
# HTTP requests for GeoIP (optional)
geoip2==4.7.0
<<<<<<< HEAD
resend==2.22.0  # Resend email service integration
=======
>>>>>>> 2b6c4e0f3e2bdec671123c59cab390bd0dde93d7
